Extract circuit description to its own source file
